requirements.txt
attrs==22.1.0
certifi==2022.6.15
charset-normalizer==2.1.1
Cython==0.29.32
greenlet==3.2.3
idna==3.3
iniconfig==1.1.1
numpy==2.3.1
packaging==21.3
pandas==2.3.1
pluggy==1.0.0
procyclingstats==0.2.5
psycopg==3.2.9
psycopg2-binary==2.9.10
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-subtests==0.8.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.28.1
selectolax==0.3.31
six==1.17.0
SQLAlchemy==2.0.41
tabulate==0.8.10
tomli==2.0.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.12
